# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * ShopNow Electronics API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: number;
  name: string;
  brand: string;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  /** @nullable */
  discountPct?: number | null;
  category: string;
  imageUrl: string;
  rating: number;
  reviewCount: number;
  inStock: boolean;
  /** @nullable */
  stockCount?: number | null;
  /** @nullable */
  specs?: string | null;
  isFeatured?: boolean;
  isDeal?: boolean;
}

export type RecommendationWidgetType = typeof RecommendationWidgetType[keyof typeof RecommendationWidgetType];


export const RecommendationWidgetType = {
  content_based: 'content_based',
  collaborative: 'collaborative',
  hybrid: 'hybrid',
} as const;

export interface RecommendedProduct {
  product: Product;
  reason: string;
}

export interface RecommendationWidget {
  type: RecommendationWidgetType;
  title: string;
  subtitle: string;
  products: RecommendedProduct[];
}

export interface HomepageRecommendations {
  contentBased: RecommendationWidget;
  collaborative: RecommendationWidget;
  hybrid: RecommendationWidget;
}

export interface PdpRecommendations {
  frequentlyBoughtTogether: RecommendationWidget;
  contentBased: RecommendationWidget;
  hybrid: RecommendationWidget;
}

export interface CartRecommendations {
  crossSell: RecommendationWidget;
  collaborative: RecommendationWidget;
  hybrid: RecommendationWidget;
}

export interface CartItem {
  product: Product;
  quantity: number;
}

export interface Cart {
  items: CartItem[];
  subtotal: number;
  discount: number;
  deliveryFee: number;
  total: number;
  /** @nullable */
  couponApplied: string | null;
}

export interface CartItemInput {
  productId: number;
  quantity: number;
}

export interface CartItemUpdate {
  quantity: number;
}

export interface User {
  id: number;
  name: string;
  email: string;
}

export interface RegisterInput {
  name: string;
  email: string;
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface AuthMessage {
  message: string;
}

export type CheckoutInputAddress = {
  name?: string;
  street?: string;
  city?: string;
  zip?: string;
};

export type CheckoutInputPayment = {
  cardNumber?: string;
  expiry?: string;
  cvv?: string;
};

export interface CheckoutInput {
  address: CheckoutInputAddress;
  payment: CheckoutInputPayment;
}

export interface OrderSuccess {
  message: string;
  orderId: number;
}

export type OrderShippingAddress = { [key: string]: unknown };

export type OrderPaymentDetails = { [key: string]: unknown };

export interface Order {
  id: number;
  totalAmount: string;
  status: string;
  createdAt: string;
  shippingAddress?: OrderShippingAddress;
  paymentDetails?: OrderPaymentDetails;
}

export type AIChatInputHistoryItem = {
  role?: string;
  content?: string;
};

export interface AIChatInput {
  message: string;
  history?: AIChatInputHistoryItem[];
}

export interface AIChatResponse {
  response: string;
}

export interface ProductSearchResult {
  products: Product[];
  total: number;
  page: number;
  totalPages: number;
}

export interface CategoryInfo {
  name: string;
  count: number;
}

export type ListProductsParams = {
category?: string;
featured?: boolean;
limit?: number;
};

export type SearchProductsParams = {
/**
 * Text search query (matches name, brand, category)
 */
q?: string;
category?: string;
minPrice?: number;
maxPrice?: number;
inStock?: boolean;
sortBy?: SearchProductsSortBy;
page?: number;
limit?: number;
};

export type SearchProductsSortBy = typeof SearchProductsSortBy[keyof typeof SearchProductsSortBy];


export const SearchProductsSortBy = {
  relevance: 'relevance',
  price_asc: 'price_asc',
  price_desc: 'price_desc',
  rating: 'rating',
  newest: 'newest',
} as const;

export type GetHomepageRecommendationsParams = {
/**
 * @nullable
 */
userId?: number | null;
};