# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ShopNow Electronics API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "200":
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Validation error or email already exists
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get currently authenticated user profile
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products with optional filters
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: featured
          in: query
          schema:
            type: boolean
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
...
  /products/search:
    get:
      operationId: searchProducts
      tags: [products]
      summary: Search products with text query and filters
      parameters:
        - name: q
          in: query
          schema:
            type: string
          description: Text search query (matches name, brand, category)
        - name: category
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: inStock
          in: query
          schema:
            type: boolean
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [relevance, price_asc, price_desc, rating, newest]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Search results with pagination
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductSearchResult"
...
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
...
  /cart/items:
    post:
      operationId: addToCart
      tags: [cart]
      summary: Add a product to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemInput"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /cart/items/{productId}:
    patch:
      operationId: updateCartItem
      tags: [cart]
      summary: Update quantity of a cart item
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemUpdate"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: removeFromCart
      tags: [cart]
      summary: Remove item from cart
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
...
components:
  schemas:
...
    Product:
      type: object
      required:
        - id
        - name
        - brand
        - price
        - category
        - imageUrl
        - rating
        - reviewCount
        - inStock
      properties:
        id:
          type: integer
        name:
          type: string
        brand:
          type: string
        price:
          type: number
        originalPrice:
          type: ["number", "null"]
        discountPct:
          type: ["integer", "null"]
        category:
          type: string
        imageUrl:
          type: string
        rating:
          type: number
        reviewCount:
          type: integer
        inStock:
          type: boolean
        stockCount:
          type: ["integer", "null"]
        specs:
          type: ["string", "null"]
        isFeatured:
          type: boolean
        isDeal:
          type: boolean
...
    CartItem:
      type: object
      required:
        - product
        - quantity
      properties:
        product:
          $ref: "#/components/schemas/Product"
        quantity:
          type: integer
    Cart:
      type: object
      required:
        - items
        - subtotal
        - discount
        - deliveryFee
        - total
        - couponApplied
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        subtotal:
          type: number
        discount:
          type: number
        deliveryFee:
          type: number
        total:
          type: number
        couponApplied:
          type: ["string", "null"]
    CartItemInput:
      type: object
      required:
        - productId
        - quantity
      properties:
        productId:
          type: integer
        quantity:
          type: integer
    CartItemUpdate:
      type: object
      required:
        - quantity
      properties:
        quantity:
          type: integer
    User:
      type: object
      required:
        - id
        - name
        - email
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
    RegisterInput:
      type: object
      required:
        - name
        - email
        - password
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
    LoginInput:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
        password:
          type: string
...
    ProductSearchResult:
      type: object
      required:
        - products
        - total
        - page
        - totalPages
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
...